Refactored Structure
