--- src/components/ProductList.css
@@ -1,5 +1,0 @@
-.card img{
-  width: 200px;
-  height: 200px;
-  margin: 0 auto;
-}
