+++ src/containers/ProductList/ProductList.css
@@ -1,0 +1,5 @@
+.card img{
+    width: 200px;
+    height: 200px;
+    margin: 0 auto;
+  }
